requirements.txt
customtkinter
fpdf2
Pillow
